Rename pkg/body-based-routing to pkg/bbr (#664)
