+++ test/integration/bbr/hermetic_test.go
@@ -29,7 +29,7 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/protobuf/testing/protocmp"
-	runserver "sigs.k8s.io/gateway-api-inference-extension/pkg/body-based-routing/server"
+	runserver "sigs.k8s.io/gateway-api-inference-extension/pkg/bbr/server"
 	logutil "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/util/logging"
 	integrationutils "sigs.k8s.io/gateway-api-inference-extension/test/integration"
 )
